requirements.txt
openai>=1.3.0
websockets>=11.0.3
python-dotenv>=1.0.0
requests>=2.31.0
sounddevice>=0.4.6
numpy>=1.24.3
scipy>=1.11.3